Moved KthGreatest into lucene.search package
