+++ elastiknn-lucene/src/main/java/org/apache/lucene/search/KthGreatest.java
@@ -1,4 +1,4 @@
-package com.klibisz.elastiknn.models;
+package org.apache.lucene.search;
 
 public class KthGreatest {
     /**
