+++ elastiknn-models/src/main/java/com/klibisz/elastiknn/models/ExactModel.java
@@ -2,6 +2,8 @@ package com.klibisz.elastiknn.models;
 
 import java.util.Arrays;
 
+import static com.klibisz.elastiknn.models.Utils.sortedIntersectionCount;
+
 public class ExactModel {
 
     @FunctionalInterface
@@ -12,34 +14,6 @@ public class ExactModel {
     @FunctionalInterface
     public interface SparseBool {
         double similarity(int[] v1, int[] v2, int totalIndices);
-    }
-
-    /**
-     * Compute the number of intersecting (i.e. identical) elements between two int arrays.
-     * IMPORTANT: Assumes the given arrays are already sorted in ascending order and _does not_ check if this is true.
-     * If the given arrays are not sorted, the answer will be wrong.
-     * This is implemented in Java because for some reason Scala will Box the ints in some cases which is unnecessary
-     * and far slower.
-     * @param xs
-     * @param ys
-     * @return The number of identical elements in the two arrays. For example {1,2,3}, {2,3,4} would return 2.
-     */
-    public static int sortedIntersectionCount(final int [] xs, final int [] ys) {
-        int n = 0;
-        int xi = 0;
-        int yi = 0;
-        while (xi < xs.length && yi < ys.length) {
-            int x = xs[xi];
-            int y = ys[yi];
-            if (x < y) xi += 1;
-            else if (x > y) yi += 1;
-            else {
-                n += 1;
-                xi += 1;
-                yi += 1;
-            }
-        }
-        return n;
     }
 
     public static class Jaccard implements SparseBool {
