+++ elastiknn-models/src/main/java/com/klibisz/elastiknn/models/HashingModel.java
@@ -10,14 +10,6 @@ public class HashingModel {
 
     public interface DenseFloat {
         HashAndFreq[] hash(float[] values);
-
-        static float dot(float[] v1, float[] v2) {
-            float dp = 0f;
-            for (int i = 0; i < v1.length; i++) dp += v1[i] * v2[i];
-            return dp;
-        }
-
     }
-
 
 }
